+++ HISTORY.md
@@ -62,6 +62,9 @@
     * German
     * Greek, thank you, HM100!
 
+### 2022-08-01
+* blocks, objects, store: include a representation of the stage as sprite in the "scripts pic" export if it is more than a single one
+
 ### 2022-07-31
 * threads: fixed #3085 (I hope ^^)
 * byob: fixed #3088 (I hope ^^)